fix: gap_hunt gh_ctx mpz leak/UB (init+clear g.b0/g.P) and halfclass cap_off uint32_t truncation
gap_hunt.c: g.b0/g.P were used via mpz_set after only memset(0) (no mpz_init;
relied on GMP default realloc(NULL)) and never cleared (mpz_clears listed only
bk,wb) -> one-time per-run leak. Now mpz_init both and clear all four.
halfclass.c: halfclass_collect_hidden_candidates_impl cap_off had a stray
(uint32_t) cast truncating the uint64_t capacity (latent overflow). Removed.
Build clean (WITH_CUDA=1 WITH_CGBN_FERMAT=1); test_halfclass PASS; gap_hunt
KMAX smoke rc=0.

+++ new_src/halfclass.c
@@ -176,9 +176,9 @@ static int64_t halfclass_collect_hidden_candidates_impl(
        no candidate is ever silently dropped (a dropped interior prime
        would fabricate a false gap). */
     uint64_t cap_off =
-        (uint32_t)((interval / 60ULL) *
-                       (uint64_t)__builtin_popcountll(halfclass_hidden_mask()) +
-                   16ULL);
+        (interval / 60ULL) *
+            (uint64_t)__builtin_popcountll(halfclass_hidden_mask()) +
+        16ULL;
     uint64_t *cand = (uint64_t *)malloc((size_t)cap_off * sizeof(uint64_t));
     if (!cand) {
         free(bits);
